1.10.0
- Add/Rebalance Biomes
- Rebalance experience levels
- Rebalance happiness
- Rebalance random events
- Added Lumber Mill
- Remove unneeded code for backpack heads
- Added sub-command to backpack for GUI instead of physical item
- Remove unused dbg cmds
- Fix armor component wanting strength instead of resistance
- Cleaned some code with dropping items
- Fix ores not dropping XP
- Fix potion effects on randomTP not going away
- Fix some econ troubles
- Fix labs not getting beakers
- Fix structure placing inside of player instead of in front
- Change permissions for war respawn signs useages
- Fix Market looking for IDs instead of getting all available items
- Fix getting baseValue of health instead of actual value
- Fix cobwebs being existant at villager's feets when structures finish
- Change message for structure damage w/ enhancements
- Moved structures to their own timer from the UpdateEventTimer

+++ civcraft/src/com/avrgaming/civcraft/structure/Bank.java
@@ -136,7 +136,7 @@ public class Bank extends Structure {
 					+CivColor.ITALIC+"(With this buff, our interest rate went from "+interestRate+"% to "+effectiveInterestRate+"%.)");
 		}
 		
-		int newCoins = (int)(principal*effectiveInterestRate);
+		double newCoins = principal*effectiveInterestRate;
 		if (newCoins != 0) {
 			CivMessage.sendTown(this.getTown(), CivColor.LightGreen+"Our town earned "+newCoins+" coins from an interest rate of "
 					+effectiveInterestRate+"% and on a principal of "+principal+" coins.");
